refactor(livechat): export the Chat connector as the route default
Make the Chat route's default export the connector, import the container
directly in the connector instead of via the barrel, and update App to
render the default Chat export.
