+++ packages/livechat/src/routes/Chat/index.ts
@@ -1,3 +1,2 @@
 export { default as Chat } from './component';
-export { default as ChatContainer } from './container';
-export { default as ChatConnector } from './connector';
+export { default } from './connector';
